+++ MuvrKit/MKScalarPredictor.swift
@@ -19,26 +19,12 @@ public protocol MKScalarPredictor {
     func trainPositional(trainingSet: [Double], forExerciseId exerciseId: MKExercise.Id)
     
     ///
-    /// Returns the scalar prediction for the ``n`` the instance of the ``exerciseId``
+    /// Returns the scalar prediction for the ``exerciseId``
     /// - parameter exerciseId: the exercise id
     /// - parameter n: the exercise number, starting at 0
     /// - returns: the predicted weight
     ///
     func predictScalarForExerciseId(exerciseId: MKExercise.Id, n: Int) -> Double?
- 
-    ///
-    /// Sets the correct prediction for exerciseId at n
-    /// - parameter exerciseId: the exercise id
-    /// - parameter n: the exercise number, starting at 0
-    /// - parameter actual: the actual value
-    ///
-    func correctScalarForExerciseId(exerciseId: MKExercise.Id, n: Int, actual: Double)
-    
-    ///
-    /// Sets the boosting function to "motivate the headcounts"
-    /// - parameter boost: the multiplier, typically close to 1.0
-    ///
-    func setBoost(boost: Float)
     
 }
 
@@ -50,7 +36,7 @@ public extension MKScalarPredictor {
             return try! NSJSONSerialization.dataWithJSONObject(self.metadata, options: [])
         }
     }
-    
+
 }
 
 enum MKScalarPredictorError: ErrorType {
